+++ DOEAssessmentApp/DOEAssessmentApp/DOE_models/functionality_model.py
@@ -9,7 +9,7 @@ class Functionality(db.Model):
     retake_assessment_days = db.Column(db.Integer, nullable=False)
     area_id = db.Column(db.Integer, nullable=False)
     proj_id = db.Column(db.Integer, nullable=False)
-    assessmentcompletion = db.Column(db.Integer)
+    assessmentcompletion = db.Column(db.Numeric(3, 2))
     achievedpercentage = db.Column(db.Numeric(3, 2))
     creationdatetime = db.Column(db.DateTime, nullable=False, server_default=func.now())
     updationdatetime = db.Column(db.DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
